Fix console errors in Signup.jsx
The two errors here were React being confused by the block and bsSize
props that are passed to LoaderButton. These may have had some effect
back when LoaderButton relied on the react-bootstrap Button. But now it
relies on the trussworks/uswds Button, for which these props do nothing.

Removing them had no visible effect on the page - which makes sense,
because if React had been able to do anything with them, it probably
wouldn't have been complaining in the console.